pyproject.toml
[project]
name = "hackathon-sia"
version = "0.1.0"
description = "Self-Improving IT Incident Agent — Google ADK + MCP + Reflexion loop"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.25.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
    "ddtrace>=4.0.0",
    "braintrust>=0.5.0",
    "autoevals>=0.1.0",
    "streamlit>=1.35.0",
    "mcp-servicenow @ git+https://github.com/jschuller/mcp-server-servicenow.git",
]

[project.optional-dependencies]
dev = ["ruff", "pytest"]

[dependency-groups]
dev = ["ruff", "pytest"]

[tool.ruff]
target-version = "py312"
line-length = 100
